// backend/src/models/enums.ts
export enum MatchState {
  OPEN = "OPEN",
  RUNNING = "RUNNING",
  COMPLETE = "COMPLETE",
}

export enum CurrencyType {
  NONE = "NONE",
  SOFT = "SOFT",
  HARD = "HARD",
}

export enum GameModeType {
  DEFAULT = "DEFAULT",
}

export enum TransactionType {
  ENTRY_FEE = "ENTRY_FEE",
  REWARD_CLAIM = "REWARD_CLAIM",
}